Refactored the particle solver classes and moved them to a different directory.
