+++ pixi/src/main/java/org/openpixi/pixi/physics/movement/solver/EmptyParticleSolver.java
@@ -1,9 +1,9 @@
-package org.openpixi.pixi.physics.solver;
+package org.openpixi.pixi.physics.movement.solver;
 
 import org.openpixi.pixi.physics.force.Force;
 import org.openpixi.pixi.physics.particles.IParticle;
 
-public class EmptyParticleSolver implements Solver
+public class EmptyParticleSolver implements ParticleSolver
 {
     public EmptyParticleSolver()
     {
